// src/action-types/action-types.ts
enum ActionTypes {
    ADD_TO_TASKS = 'ADD_TO_TASKS',
    DELETE_OF_TASKS = ' DELETE_OF_TASKS',
    EDIT_TASK = 'EDIT_TASK',
    TOGGLE_TASK = 'TOGGLE_TASK',
    HOVER_TASK = 'HOVER_TASK',
    CLEAR_ALL_TASK = 'CLEAR_ALL_TASK',
    FETCH_WEATHER_LOAD = 'FETCH_WEATHER_LOAD',
    FETCH_WEATHER_SUCCESS = 'FETCH_WEATHER_SUCCESS',
    FETCH_WEATHER_FAILURE = 'FETCH_WEATHER_FAILURE',
    FETCH_INPUT_SEARCH = 'FETCH_INPUT_SEARCH',
    FETCH_CITY_SUCCESS = 'FETCH_CITY_SUCCESS',
    FETCH_CITY_FAILURE = 'FETCH_CITY_FAILURE',
    FETCH_CITY_LOADING = ' FETCH_CITY_LOADING'
}

export default ActionTypes